+++ src/main/java/ru/mail/polis/ads/vadim01er/part4/Task3.java
@@ -1,4 +1,4 @@
-package ru.mail.polis.ads.part4.vadim01er;
+package ru.mail.polis.ads.vadim01er.part4;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -6,24 +6,27 @@ import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.util.StringTokenizer;
 
-// https://www.e-olymp.com/ru/submissions/7526272
+// https://www.e-olymp.com/ru/submissions/7525773
 
-public class Task4 {
+public class Task3 {
 
     public static void main(String[] args) {
         FastScanner scanner = new FastScanner(System.in);
         int n = scanner.nextInt();
-        int[][] arr = new int[2][n + 2];
-        for (int i = 1; i < n + 1; i++) arr[0][i] = scanner.nextInt();
-        int k = scanner.nextInt();
-        for (int i = 1; i < n + 2; i++) {
-            arr[1][i] = arr[1][i - 1];
-            for (int j = Math.max(i - k, 0); j < i; j++) {
-                arr[1][i] = Math.max(arr[1][i], arr[1][j]);
+        int[] firstArr = new int[n];
+        for (int i = 0; i < n; i++) firstArr[i] = scanner.nextInt();
+        int m = scanner.nextInt();
+        int[] secondArr = new int[m];
+        for (int i = 0; i < m; i++) secondArr[i] = scanner.nextInt();
+        int[][] matrix = new int[n + 1][m + 1];
+        for (int i = 0; i < n + 1; i++) {
+            for (int j = 0; j < m + 1; j++) {
+                if (i == 0 || j == 0) matrix[i][j] = 0;
+                else if (firstArr[i - 1] == secondArr[j - 1]) matrix[i][j] = matrix[i - 1][j - 1] + 1;
+                else matrix[i][j] = Math.max(matrix[i - 1][j], matrix[i][j - 1]);
             }
-            arr[1][i] += arr[0][i];
         }
-        System.out.println(arr[1][n + 1]);
+        System.out.println(matrix[n][m]);
     }
 
     private static class FastScanner {
